Change state to unsigned int.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@37482 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e_dbus/src/lib/nm/e_nm.c
@@ -46,7 +46,7 @@ cb_signal_manager_state_changed(void *data, DBusMessage *msg)
   }
 
   if (ctx->cb_manager_state_changed)
-    ctx->cb_manager_state_changed(ctx->data_manager_state_changed, (int)state);
+    ctx->cb_manager_state_changed(ctx->data_manager_state_changed, state);
 }
 
 static void
